examples/sample-ml-app/requirements.txt
kfp==2.4.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
google-cloud-storage==2.10.0
google-cloud-bigquery==3.14.1
requests==2.31.0